requirements.txt
openai
requests
Pillow
sentence-transformers
faiss-cpu
streamlit
pandas
numpy
tqdm
dotenv
pickle
json